test: enable DocumentManager-integ-test
